pyproject.toml
[project]
name = "intelligent-agent"
version = "0.1.0"
description = "Intelligent Autonomous AI Agent with LangGraph orchestration"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0.0",
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    "openai>=1.0.0",
    "anthropic>=0.39.0",
    "pydantic>=2.0",
    "structlog>=24.0.0",
    "aiosqlite>=0.20.0",
    "duckduckgo-search>=6.0.0",
    "RestrictedPython>=6.0",
    "pyyaml>=6.0",
    "python-telegram-bot>=21.0",
    "chromadb>=0.4.0",
    "requests>=2.28.0",
    "prometheus-client>=0.19.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "apscheduler>=3.10.0",
]

[project.scripts]
agent-cli = "src.interfaces.cli:run_cli"
agent-telegram = "src.interfaces.telegram_bot:run_telegram_bot"
agent-api = "src.interfaces.api:run_api"


[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=4.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]
